Create a locatable base controller for better reuse of location parameters for multi-edit
